+++ mrp_bom_line_formula_template/models/__init__.py
@@ -1,0 +1,2 @@
+from . import mrp_bom_line_formula_template
+from . import mrp_bom_line