+++ mrp_bom_line_formula_wizard/__init__.py
@@ -1,0 +1,5 @@
+# Copyright 2026 Rosen Vladimirov <vladimirov.rosen@gmail.com>
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+
+from . import models
+from . import wizard